update documentation and consolidate into helpers
